// src/components/Filters/Period/hooks/period.types.ts
export enum Period {
  perMinute = 'perMinute',
  per30Minute = 'per30minute',
  perHour = 'perHour',
  per4Hour = 'per4Hour',
  per8Hour = 'per8Hour',
  per12Hour = 'per12Hour',
  last24hour = 'last24hour',
  perDay = 'perDay',
  perYesterday = 'perYesterday',
  perWeek = 'perWeek',
  last7days = 'last7days',
  perMonth = 'perMonth',
  last30days = 'last30days',
  last6month = 'last6month',
  last3month = 'last3month',
  perAllTime = 'perAllTime',
  custom = 'custom',
}